vector_database/requirements.txt
openai-agents==0.0.14
openai==1.77.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.2.1
numpy>=1.26.4 
openpyxl>=3.1.2